# src/gen/java/io/swagger/model/Magnitude.java
/*
 * GeoCardioBand Graphics
 * API encargada de la gestión de gráficas del proyecto GeoCardioBand.
 *
 * OpenAPI spec version: 1.0.0
 * Contact: lutevi04@alumnos.unex.es
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

package io.swagger.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonValue;
import javax.validation.constraints.*;

import com.fasterxml.jackson.annotation.JsonCreator;

/**
 * Gets or Sets Magnitude
 */
public enum Magnitude {
  SPO2("spo2"),
    BPM("bpm");

  private String value;

  Magnitude(String value) {
    this.value = value;
  }

  @Override
  public String toString() {
    return String.valueOf(value);
  }

  @JsonCreator
  public static Magnitude fromValue(String text) {
    for (Magnitude b : Magnitude.values()) {
      if (String.valueOf(b.value).equals(text)) {
        return b;
      }
    }
    return null;
  }
}